Split out the tests that use conda from tests/tests.py.
The file is too big in any case, and this seems a natural
split as the conda tests are so slow.
