+++ tests/tests.py
@@ -12,7 +12,6 @@ import tempfile
 from unittest.mock import AsyncMock, patch
 
 import pytest
-from snakemake.deployment.conda import get_env_setup_done_flag_file
 from snakemake.persistence import Persistence
 from snakemake.resources import DefaultResources, GroupResources
 from snakemake.settings.enums import RerunTrigger
@@ -23,7 +22,7 @@ from snakemake.shell import shell
 
 sys.path.insert(0, os.path.dirname(__file__))
 
-from .common import run, dpath, apptainer, conda, connected
+from .common import run, dpath, apptainer, connected
 from .conftest import skip_on_windows, only_on_windows, ON_WINDOWS, needs_strace
 
 from snakemake_interface_executor_plugins.settings import (
@@ -409,28 +408,8 @@ def test_empty_include():
     run(dpath("test_empty_include"))
 
 
-@skip_on_windows
-@conda
-def test_script():
-    run(
-        dpath("test_script"),
-        deployment_method={DeploymentMethod.CONDA},
-        check_md5=False,
-    )
-
-
 def test_script_python():
     run(dpath("test_script_py"))
-
-
-@skip_on_windows
-@conda
-def test_script_rs():
-    run(
-        dpath("test_script_rs"),
-        deployment_method={DeploymentMethod.CONDA},
-        check_md5=False,
-    )
 
 
 @skip_on_windows  # Test relies on perl
@@ -507,137 +486,6 @@ def test_symlink_time_handling():
 @skip_on_windows
 def test_protected_symlink_output():
     run(dpath("test_protected_symlink_output"))
-
-
-@conda
-def test_conda():
-    run(dpath("test_conda"), deployment_method={DeploymentMethod.CONDA})
-
-
-@conda
-def test_conda_list_envs():
-    run(dpath("test_conda"), conda_list_envs=True, check_results=False)
-
-
-# TODO failing with FAILED tests/tests.py::test_conda_create_envs_only -
-# PermissionError: [WinError 32] The process cannot access the file because
-# it is being used by another process:
-# 'C:\\Users\\RUNNER~1\\AppData\\Local\\Temp\\snakemake-2q4osog0\\test-env.yaml'
-@skip_on_windows
-@conda
-def test_conda_create_envs_only():
-    tmpdir = run(
-        dpath("test_conda"),
-        conda_create_envs=True,
-        check_results=False,
-        cleanup=False,
-        cleanup_scripts=False,
-    )
-    env_dir = next(
-        (p for p in Path(tmpdir, ".snakemake", "conda").iterdir() if p.is_dir()), None
-    )
-    assert env_dir is not None
-    assert get_env_setup_done_flag_file(Path(env_dir)).exists()
-    shutil.rmtree(tmpdir)
-
-
-@conda
-def test_upstream_conda():
-    run(
-        dpath("test_conda"),
-        deployment_method={DeploymentMethod.CONDA},
-        conda_frontend="conda",
-    )
-
-
-@skip_on_windows
-@conda
-def test_deploy_script():
-    run(dpath("test_deploy_script"), deployment_method={DeploymentMethod.CONDA})
-
-
-@skip_on_windows
-@conda
-def test_deploy_hashing():
-    tmpdir = run(
-        dpath("test_deploy_hashing"),
-        deployment_method={DeploymentMethod.CONDA},
-        cleanup=False,
-    )
-    assert len(next(os.walk(os.path.join(tmpdir, ".snakemake/conda")))[1]) == 2
-
-
-@conda
-def test_conda_custom_prefix():
-    run(
-        dpath("test_conda_custom_prefix"),
-        deployment_method={DeploymentMethod.CONDA},
-        conda_prefix="custom",
-        set_pythonpath=False,
-    )
-
-
-@only_on_windows
-@conda
-def test_conda_cmd_exe():
-    # Tests the conda environment activation when cmd.exe
-    # is used as the shell
-    run(dpath("test_conda_cmd_exe"), deployment_method={DeploymentMethod.CONDA})
-
-
-@skip_on_windows  # wrappers are for linux and macos only
-@conda
-def test_wrapper():
-    run(
-        dpath("test_wrapper"),
-        deployment_method={DeploymentMethod.CONDA},
-        check_md5=False,
-    )
-
-
-@skip_on_windows  # wrappers are for linux and macos only
-@conda
-def test_wrapper_local_git_prefix():
-    import git
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        print("Cloning wrapper repo...")
-        repo = git.Repo.clone_from(
-            "https://github.com/snakemake/snakemake-wrappers", tmpdir
-        )
-        print("Cloning complete.")
-
-        run(
-            dpath("test_wrapper"),
-            deployment_method={DeploymentMethod.CONDA},
-            wrapper_prefix=f"git+file://{tmpdir}",
-            check_md5=False,
-        )
-
-
-@conda
-def test_get_log_none():
-    run(dpath("test_get_log_none"))
-
-
-@conda
-def test_get_log_both():
-    run(dpath("test_get_log_both"))
-
-
-@conda
-def test_get_log_stderr():
-    run(dpath("test_get_log_stderr"))
-
-
-@conda
-def test_get_log_stdout():
-    run(dpath("test_get_log_stdout"))
-
-
-@conda
-def test_get_log_complex():
-    run(dpath("test_get_log_complex"))
 
 
 def test_spaces_in_fnames():
@@ -873,18 +721,6 @@ def test_singularity_module_invalid():
 @skip_on_windows
 @apptainer
 @connected
-@conda
-def test_singularity_conda():
-    run(
-        dpath("test_singularity_conda"),
-        deployment_method={DeploymentMethod.CONDA, DeploymentMethod.APPTAINER},
-        conda_frontend="conda",
-    )
-
-
-@skip_on_windows
-@apptainer
-@connected
 def test_singularity_none():
     run(dpath("test_singularity_none"), deployment_method={DeploymentMethod.APPTAINER})
 
@@ -908,11 +744,6 @@ def test_bash():
 
 def test_inoutput_is_path():
     run(dpath("test_inoutput_is_path"))
-
-
-@conda
-def test_archive():
-    run(dpath("test_archive"), archive="workflow-archive.tar.gz")
 
 
 def test_log_input():
@@ -1304,16 +1135,6 @@ def test_issue930():
     run(dpath("test_issue930"), cluster="./qsub")
 
 
-@skip_on_windows
-@conda
-def test_issue635():
-    run(
-        dpath("test_issue635"),
-        deployment_method={DeploymentMethod.CONDA},
-        check_md5=False,
-    )
-
-
 # TODO remove skip
 @pytest.mark.skip(
     reason="Temporarily disable until the stable container image becomes available again."
@@ -1338,11 +1159,6 @@ def test_issue1037():
     )
 
 
-@conda
-def test_issue1046():
-    run(dpath("test_issue1046"))
-
-
 def test_checkpoints():
     run(dpath("test_checkpoints"))
 
@@ -1353,12 +1169,6 @@ def test_checkpoints_dir():
 
 def test_issue1092():
     run(dpath("test_issue1092"))
-
-
-@skip_on_windows
-@conda
-def test_issue1093():
-    run(dpath("test_issue1093"), deployment_method={DeploymentMethod.CONDA})
 
 
 def test_issue958():
@@ -1563,25 +1373,6 @@ def test_string_resources():
         cluster="./qsub.py",
     )
 
-
-@conda
-def test_jupyter_notebook():
-    run(dpath("test_jupyter_notebook"), deployment_method={DeploymentMethod.CONDA})
-
-
-@conda
-def test_jupyter_notebook_draft():
-    from snakemake.settings.types import NotebookEditMode
-
-    run(
-        dpath("test_jupyter_notebook_draft"),
-        deployment_method={DeploymentMethod.CONDA},
-        edit_notebook=NotebookEditMode(draft_only=True),
-        targets=["results/result_intermediate.txt"],
-        check_md5=False,
-    )
-
-
 def test_github_issue456():
     run(dpath("test_github_issue456"))
 
@@ -1630,22 +1421,6 @@ def test_paramspace_single_wildcard():
 
 def test_github_issue806():
     run(dpath("test_github_issue806"), config=dict(src_lang="es", trg_lang="en"))
-
-
-@skip_on_windows
-@apptainer
-@conda
-def test_containerized():
-    run(
-        dpath("test_containerized"),
-        deployment_method={DeploymentMethod.CONDA, DeploymentMethod.APPTAINER},
-    )
-
-
-@skip_on_windows
-@conda
-def test_containerize():
-    run(dpath("test_conda"), containerize=True, check_results=False)
 
 
 def test_long_shell():
@@ -1816,15 +1591,6 @@ def test_github_issue1158():
     run(dpath("test_github_issue1158"), cluster="./qsub.py")
 
 
-@conda
-def test_converting_path_for_r_script():
-    run(
-        dpath("test_converting_path_for_r_script"),
-        cores=1,
-        deployment_method={DeploymentMethod.CONDA},
-    )
-
-
 def test_ancient_dag():
     run(dpath("test_ancient_dag"))
 
@@ -1844,22 +1610,6 @@ def test_issue1331():
     # not guaranteed to fail, so let's try multiple times
     for _ in range(10):
         run(dpath("test_issue1331"), cores=4)
-
-
-@skip_on_windows
-@conda
-def test_conda_named():
-    run(dpath("test_conda_named"), deployment_method={DeploymentMethod.CONDA})
-
-
-@skip_on_windows
-@conda
-def test_conda_function():
-    run(
-        dpath("test_conda_function"),
-        deployment_method={DeploymentMethod.CONDA},
-        cores=1,
-    )
 
 
 @skip_on_windows
@@ -2018,37 +1768,9 @@ def test_module_input_func():
     run(dpath("test_module_input_func"))
 
 
-@skip_on_windows  # the testcase only has a linux-64 pin file
-@conda
-def test_conda_pin_file():
-    run(dpath("test_conda_pin_file"), deployment_method={DeploymentMethod.CONDA})
-
-
 @skip_on_windows  # sufficient to test this on linux
 def test_github_issue1618():
     run(dpath("test_github_issue1618"), cores=5)
-
-
-@conda
-def test_conda_python_script():
-    run(dpath("test_conda_python_script"), deployment_method={DeploymentMethod.CONDA})
-
-
-@conda
-def test_conda_python_3_7_script():
-    run(
-        dpath("test_conda_python_3_7_script"),
-        deployment_method={DeploymentMethod.CONDA},
-    )
-
-
-@conda
-def test_prebuilt_conda_script():
-    sp.run(
-        f"conda env create -f {dpath('test_prebuilt_conda_script/env.yaml')}",
-        shell=True,
-    )
-    run(dpath("test_prebuilt_conda_script"), deployment_method={DeploymentMethod.CONDA})
 
 
 @skip_on_windows
@@ -2136,23 +1858,8 @@ def test_load_metawrapper():
     run(dpath("test_load_metawrapper"), executor="dryrun")
 
 
-@skip_on_windows
-@conda
-def test_conda_global():
-    run(
-        dpath("test_conda_global"),
-        deployment_method={DeploymentMethod.CONDA},
-        executor="dryrun",
-    )
-
-
 def test_missing_file_dryrun():
     run(dpath("test_missing_file_dryrun"), executor="dryrun", shouldfail=True)
-
-
-@conda
-def test_script_pre_py39():
-    run(dpath("test_script_pre_py39"), deployment_method={DeploymentMethod.CONDA})
 
 
 def test_issue1256():
@@ -2184,18 +1891,6 @@ def test_issue2574():
     assert p.returncode == 1
     assert "KeyError" in stderr
     assert "line 4," in stderr
-
-
-@conda
-def test_resource_string_in_cli_or_profile():
-    test_path = dpath("test_resource_string_in_cli_or_profile")
-    profile = os.path.join(test_path, "profiles")
-    # workflow profile is loaded by default
-    run(
-        test_path,
-        snakefile="Snakefile",
-        shellcmd=f"snakemake --workflow-profile {profile} -c1 --default-resources slurm_account=foo other_resource='--test'",
-    )
 
 
 @skip_on_windows  # not platform dependent, only cli
@@ -2317,7 +2012,7 @@ def test_github_issue2732():
 
 @skip_on_windows
 @apptainer
-def test_shell_exec():
+def test_shell_exec_singularity():
     run(dpath("test_shell_exec"), deployment_method={DeploymentMethod.APPTAINER})
